move frontend files around
