--- frontend/src/pages/index.ts
@@ -1,3 +1,0 @@
-export { Home } from "./Home";
-export { Apps } from "./Apps";
-export { AppChat } from "./AppChat";
