--- frontend/src/types/index.ts
@@ -1,3 +1,0 @@
-// Re-export all types from different domains
-export * from "./api";
-export * from "./files";
